requirements.txt
lief>=0.13.0
Pillow>=10.0.0
numpy>=1.24.0
streamlit>=1.28.0
capstone>=5.0.0
pytest>=7.4.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
